Refactor microphone sound level functions and remove deprecated soundLevel shim
